[Graphics] Resolve the issue of game losing response and no graphics rendered
SDL requires a loop of continuous polling events, while the previous
version of code does not contain the code, which resulted in losing
response.

The current version fixed the problem, but does not fix other bugs of
the game. There is also no guarantee that other parts work. This fix
also fixed the problem that no graphics is drawn. This is because
rendererPresent is never called